JS to prevent users from closing the option-select
One of the side effects of the closing link was that it served as an
indication that the list of checkboxes had a checked box when the
option select is closed.

This is a useful side effect from it's intended purpose.

The proposed fix here is to prevent the user from closing the widget
if they have selected an option.